+++ packages/vuetify/src/components/VCheckbox/VCheckbox.tsx
@@ -3,11 +3,7 @@ import './VCheckbox.sass'
 
 // Components
 import { filterInputProps, makeVInputProps, VInput } from '@/components/VInput/VInput'
-import { filterControlProps, makeSelectionControlProps } from '@/components/VSelectionControl/VSelectionControl'
-import { VCheckboxBtn } from './VCheckboxBtn'
-
-// Composables
-import { useProxiedModel } from '@/composables/proxiedModel'
+import { filterCheckboxBtnProps, makeVCheckboxBtnProps, VCheckboxBtn } from './VCheckboxBtn'
 
 // Utility
 import { defineComponent, filterInputAttrs, useRender } from '@/util'
@@ -18,42 +14,15 @@ export const VCheckbox = defineComponent({
   inheritAttrs: false,
 
   props: {
-    indeterminate: Boolean,
-    indeterminateIcon: {
-      type: String,
-      default: '$checkboxIndeterminate',
-    },
-
     ...makeVInputProps(),
-    ...makeSelectionControlProps(),
-
-    falseIcon: {
-      type: String,
-      default: '$checkboxOff',
-    },
-    trueIcon: {
-      type: String,
-      default: '$checkboxOn',
-    },
-  },
-
-  emits: {
-    'update:indeterminate': (val: boolean) => true,
+    ...makeVCheckboxBtnProps(),
   },
 
   setup (props, { attrs, slots }) {
-    const indeterminate = useProxiedModel(props, 'indeterminate')
-
-    function onChange () {
-      if (indeterminate.value) {
-        indeterminate.value = false
-      }
-    }
-
     useRender(() => {
       const [inputAttrs, controlAttrs] = filterInputAttrs(attrs)
       const [inputProps, _1] = filterInputProps(props)
-      const [controlProps, _2] = filterControlProps(props)
+      const [checkboxProps, _2] = filterCheckboxBtnProps(props)
 
       return (
         <VInput
@@ -68,9 +37,7 @@ export const VCheckbox = defineComponent({
               isReadonly,
             }) => (
               <VCheckboxBtn
-                { ...controlProps }
-                indeterminate={ indeterminate.value }
-                onUpdate:modelValue={ onChange }
+                { ...checkboxProps }
                 disabled={ isDisabled.value }
                 readonly={ isReadonly.value }
                 { ...controlAttrs }
