+++ packages/vuetify/src/components/VCheckbox/VCheckboxBtn.tsx
@@ -1,6 +1,16 @@
-import { defineComponent, propsFactory } from '@/util'
-import { computed } from 'vue'
+// Componenst
 import { VSelectionControl } from '../VSelectionControl'
+
+// Composables
+import { useProxiedModel } from '@/composables/proxiedModel'
+
+// Utility
+import { computed } from 'vue'
+import { defineComponent, pick, propsFactory, useRender } from '@/util'
+import { makeSelectionControlProps } from '../VSelectionControl/VSelectionControl'
+
+// Types
+import type { ExtractPropTypes } from 'vue'
 
 export const makeVCheckboxBtnProps = propsFactory({
   indeterminate: Boolean,
@@ -8,6 +18,9 @@ export const makeVCheckboxBtnProps = propsFactory({
     type: String,
     default: '$checkboxIndeterminate',
   },
+
+  ...makeSelectionControlProps(),
+
   falseIcon: {
     type: String,
     default: '$checkboxOff',
@@ -16,49 +29,61 @@ export const makeVCheckboxBtnProps = propsFactory({
     type: String,
     default: '$checkboxOn',
   },
-  disabled: Boolean,
-  readonly: Boolean,
-  modelValue: null,
 })
 
 export const VCheckboxBtn = defineComponent({
   name: 'VCheckboxBtn',
 
-  inheritAttrs: false,
-
   props: makeVCheckboxBtnProps(),
 
   emits: {
-    'update:modelValue': (value: boolean) => true,
+    'update:modelValue': (value: any) => true,
+    'update:indeterminate': (val: boolean) => true,
   },
 
-  setup (props, { attrs, slots, emit }) {
+  setup (props, { slots, emit }) {
+    const indeterminate = useProxiedModel(props, 'indeterminate')
+
+    function onChange (v: any) {
+      if (indeterminate.value) {
+        indeterminate.value = false
+      }
+
+      emit('update:modelValue', v)
+    }
+
     const falseIcon = computed(() => {
       return props.indeterminate
         ? props.indeterminateIcon
         : props.falseIcon
     })
+
     const trueIcon = computed(() => {
       return props.indeterminate
         ? props.indeterminateIcon
         : props.trueIcon
     })
 
-    return () => (
+    useRender(() => (
       <VSelectionControl
-        { ...attrs }
-        class={['v-checkbox-btn', attrs.class]}
+        { ...props }
+        class="v-checkbox-btn"
         type="checkbox"
         inline
-        modelValue={ props.modelValue }
-        onUpdate:modelValue={ v => emit('update:modelValue', v) }
+        onUpdate:modelValue={ onChange }
         falseIcon={ falseIcon.value }
         trueIcon={ trueIcon.value }
         aria-checked={ props.indeterminate ? 'mixed' : undefined }
-        disabled={ props.disabled }
-        readonly={ props.readonly }
         v-slots={ slots }
       />
-    )
+    ))
+
+    return {}
   },
 })
+
+export type VCheckboxBtn = InstanceType<typeof VCheckboxBtn>
+
+export function filterCheckboxBtnProps (props: ExtractPropTypes<ReturnType<typeof makeVCheckboxBtnProps>>) {
+  return pick(props, Object.keys(VCheckboxBtn.props) as any)
+}
